Extract RobotContext to module
